SignUp error handling
Error handling for signUpUser  function; fixed few UI issues

+++ app/src/main/java/com/example/sculptify/data/User.kt
@@ -9,6 +9,6 @@ data class User(
     val weeklyGoal: Int = 3,
     val gender: String = "",
     val height: Int = 0,
-    val weight: Int = 0,
+    val weight: Float = 0f,
     val yearOfBirth: Int = 0
 )
